Small changes, renamed keymap Added MicMute Revived all dead keys Removed OS Shift
