+++ keyboards/s60-x/keymaps/ansi_qwertz/config.h
